TransformSpecifications.jl

TransformSpecifications.TransformSpecifications — Module

TransformSpecifications

This package enables structured transform elements via defined I/O specifications.

TransformSpecifications.AbstractTransformSpecification — Type

abstract type AbstractTransformSpecification

Transform specifications are represented by subtypes of AbstractTransformSpecification. Each leaf should be immutable and define methods for

input_specification returns type expected/allowed as transform input

output_specification returns output type generated by successfully completed processing

transform!, which transforms an input of type input_specification and returns an output of type output_specification.

It may additionally define a custom non-mutating transform function.

TransformSpecifications.DAGStep — Type

DAGStep

Helper struct, used to construct NoThrowDAGs. Requires fields

name::String: Name of step, must be unique across a constructed DAG

input_assembler::TransformSpecification: Transform used to construct step's input; see input_assembler for details.

transform_spec::AbstractTransformSpecification: Transform applied by step

TransformSpecifications.NoThrowDAG — Type

NoThrowDAG <: AbstractTransformSpecification
NoThrowDAG(steps::AbstractVector{DAGStep})

Transform specification constructed from a DAG of transform specification nodes (steps), such that calling transform! on the DAG iterates through the steps, first constructing that step's input from all preceding upstream step outputs and then appling that step's own transform to the constructed input.

The DAG's input_specification is that of the first step in the DAG; its output_specification is that of the last step. As the first step's input is by definition the same as the overall input to the DAG, its step.input_assembler must be nothing.

DAG construction tip

As the input to the DAG at is by definition the input to the first step in that DAG, only the first step will have access to the input directly passed in by the caller. To grant access to this top-level input to downstream tasks, construct the DAG with an initial step that is an identity transform, i.e., is_identity_no_throw_transform(first(steps)) returns true. Downstream steps can then depend on the output of specific fields from this initial step. The single argument TransformSpecification constructor creates such an identity transform.

DAG construction warning

It is the caller's responsibility to implement a DAG, and to not introduce any recursion or cycles. What will happen if you do? To quote Tom Lehrer, "well, you ask a silly question, you get a silly answer!"

Storage of intermediate values

The output of each step in the DAG is stored locally in memory for the entire lifetime of the transform operation, whether or not it is actually accessed by any later steps. Large intermediate outputs may result in unexpected memory pressure relative to function composition or even local evaluation (since they are not visible to the garbage collector).

Fields

The following fields are constructed automatically when constructing a NoThrowDAG from a vector of DAGSteps:

step_transforms::OrderedDict{String,AbstractTransformSpecification}: Ordered dictionary of processing steps

step_input_assemblers::Dict{String,TransformSpecification}: Dictionary with functions for constructing the input for each key in step_transforms as a function that takes in a Dict{String,NoThrowResult} of all upstream step_transforms results.

_step_output_fields::Dict{String,Dict{Symbol,Any}}: Internal mapping of upstream step outputs to downstream inputs, used to e.g. valdiate that the input to each step can be constructed from the outputs of the upstream steps.

Example

using Legolas: @schema, @version
using TransformSpecifications: input_assembler

@schema "example-one-var" ExampleOneVarSchema
@version ExampleOneVarSchemaV1 begin
    var::String
end

@schema "example-two-var" ExampleTwoVarSchema
@version ExampleTwoVarSchemaV1 begin
    var1::String
    var2::String
end

# Say we have three functions we want to chain together:
fn_a(x) = ExampleOneVarSchemaV1(; var=x.var * "_a")
fn_b(x) = ExampleOneVarSchemaV1(; var=x.var * "_b")
fn_c(x) = ExampleOneVarSchemaV1(; var=x.var1 * x.var2 * "_c")

# First, specify these functions as transforms: what is the specification of the
# function's input and output?
step_a_transform = NoThrowTransform(ExampleOneVarSchemaV1, ExampleOneVarSchemaV1, fn_a)
step_b_transform = NoThrowTransform(ExampleOneVarSchemaV1, ExampleOneVarSchemaV1, fn_b)
step_c_transform = NoThrowTransform(ExampleTwoVarSchemaV1, ExampleOneVarSchemaV1, fn_c)

# Next, set up the DAG between the upstream outputs into each step's input:
step_b_assembler = input_assembler(upstream -> (; var=upstream["step_a"][:var]))
step_c_assembler = input_assembler(upstream -> (; var1=upstream["step_a"][:var],
                                                var2=upstream["step_b"][:var]))
# ...note that step_a is skipped, as there are no steps upstream from it.

steps = [DAGStep("step_a", nothing, step_a_transform),
         DAGStep("step_b", step_b_assembler, step_b_transform),
         DAGStep("step_c", step_c_assembler, step_c_transform)]
dag = NoThrowDAG(steps)

# output
NoThrowDAG (ExampleOneVarSchemaV1 => ExampleOneVarSchemaV1):
  🌱  step_a: ExampleOneVarSchemaV1 => ExampleOneVarSchemaV1: `fn_a`
   ·  step_b: ExampleOneVarSchemaV1 => ExampleOneVarSchemaV1: `fn_b`
  🌷  step_c: ExampleTwoVarSchemaV1 => ExampleOneVarSchemaV1: `fn_c`

This DAG can then be applied to an input, just like a regular TransformSpecification can:

input = ExampleOneVarSchemaV1(; var="initial_str")
transform!(dag, input)

# output
NoThrowResult{ExampleOneVarSchemaV1}: Transform succeeded
  ✅ result: ExampleOneVarSchemaV1:
 :var  "initial_str_ainitial_str_a_b_c"

Similarly, this transform will fail if the input specification is violated–-but because it returns a NoThrowResult, it will fail gracefully:

# What is the input specification?
input_specification(dag)

# output
ExampleOneVarSchemaV1

transform!(dag, ExampleTwoVarSchemaV1(; var1="wrong", var2="input schema"))

# output
NoThrowResult{Missing}: Transform failed
  ❌ Input to step `step_a` doesn't conform to specification `ExampleOneVarSchemaV1`. Details: ArgumentError("Invalid value set for field `var`, expected String, got a value of type Missing (missing)")

TransformSpecifications.NoThrowResult — Type

NoThrowResult(result::T, violations, warnings) where {T}
NoThrowResult(result; violations=String[], warnings=String[])
NoThrowResult(; result=missing, violations=String[], warnings=String[])

Type that specifies the result of a transformation, indicating successful application of a transform through presence (or lack thereof) of violations. Consists of either a non-missing result (success state) or non-empty violations and type Missing (failure state).

Note that constructing a NoThrowTransform from an input result of type NoThrowTransform, e.g., NoThrowTransform(::NoThrowTransform{T}, ...), collapses down to a singleNoThrowResult{T}; any inner and outer warnings and violations fields are concatenated and returned in the resultantNoThrowResult{T}`.

See also: nothrow_succeeded

Fields

warnings::AbstractVector{<:AbstractString}: List of generated warnings that are not critical enough to be violations.

violations::AbstractVector{<:AbstractString} List of reason(s) result was not able to be generated.

result::: Generated result; missing if any violations encountered.

Example

using Legolas: @schema, @version
@schema "example" ExampleSchemaA
@version ExampleSchemaAV1 begin
    name::String
end

NoThrowResult(ExampleSchemaAV1(; name="yeehaw"))

# output
NoThrowResult{ExampleSchemaAV1}: Transform succeeded
  ✅ result: ExampleSchemaAV1:
 :name  "yeehaw"

NoThrowResult(ExampleSchemaAV1(; name="huzzah"); warnings="Hark, watch your step...")

# output
NoThrowResult{ExampleSchemaAV1}: Transform succeeded
  ⚠️  Hark, watch your step...
  ✅ result: ExampleSchemaAV1:
 :name  "huzzah"

NoThrowResult(; violations=["Epic fail!", "Slightly less epic fail!"],
                     warnings=["Uh oh..."])

# output
NoThrowResult{Missing}: Transform failed
  ❌ Epic fail!
  ❌ Slightly less epic fail!
  ⚠️  Uh oh...

TransformSpecifications.NoThrowTransform — Type

NoThrowTransform{TransformSpecification{T<:Type,U<:Type}} <: AbstractTransformSpecification

Wrapper around a basic TransformSpecification that returns a NoThrowResult of type NoThrowResult{T}, where T is the output specification of the inner transform. If calling transform! on a NoThrowTransform errors, due to either incorrect input/output types or an exception during the transform itself, the exception will be caught and returned as a NoThrowResult{Missing}, with the error(s) in the result's violations field. See NoThrowResult for details.

Note that results of a NoThrowTransform collapse down to a single NoThrowResult when nested, such that if the outputspecification of the inner TransformSpecification is itself a NoThrowResult{T}, the outputspecification of the NoThrowTransform will have that same output specification NoThrowResult{T}, and not NoThrowResult{NoThrowResult{T}}.

Fields

transform_spec::TransformSpecification{T,U}

Example 1: Successful transformation

Set-up:

using Legolas: @schema, @version

@schema "example-a" ExampleSchemaA
@version ExampleSchemaAV1 begin
    in_name::String
end

@schema "example-b" ExampleSchemaB
@version ExampleSchemaBV1 begin
    out_name::String
end

function apply_example(in_record)
    out_name = in_record.in_name * " earthling"
    return ExampleSchemaBV1(; out_name)
end
ntt = NoThrowTransform(ExampleSchemaAV1, ExampleSchemaBV1, apply_example)

# output
NoThrowTransform{ExampleSchemaAV1,ExampleSchemaBV1}: `apply_example`

Application of transform:

transform!(ntt, ExampleSchemaAV1(; in_name="greetings"))

# output
NoThrowResult{ExampleSchemaBV1}: Transform succeeded
  ✅ result: ExampleSchemaBV1:
 :out_name  "greetings earthling"

Example 2: Failing transformation

Set-up:

force_failure_example(in_record) = NoThrowResult(; violations=["womp", "womp"])
ntt = NoThrowTransform(ExampleSchemaAV1, ExampleSchemaBV1, force_failure_example)

# output
NoThrowTransform{ExampleSchemaAV1,ExampleSchemaBV1}: `force_failure_example`

Application of transform:

transform!(ntt, ExampleSchemaAV1(; in_name="greetings"))

# output
NoThrowResult{Missing}: Transform failed
  ❌ womp
  ❌ womp

TransformSpecifications.NoThrowTransform — Method

NoThrowTransform(specification::Type)

Create NoThrowTransform that meets the criteria of an identity NoThrowTransform, i.e., is_identity_no_throw_transform.

See also: identity_no_throw_result

TransformSpecifications.TransformSpecification — Type

TransformSpecification{T<:Type,U<:Type} <: AbstractTransformSpecification

Basic component that specifies a transform that, when applied to input of type T, will return output of type U.

See also: TransformSpecification

Fields

input_specification::T

output_specification::U

transform_fn::Function Function with signature transform_fn(::input_specification) -> output_specification

Example

using Legolas: @schema, @version

@schema "example-in" ExampleInSchema
@version ExampleInSchemaV1 begin
    in_name::String
end

@schema "example-out" ExampleOutSchema
@version ExampleOutSchemaV1 begin
    out_name::String
end

function apply_example(in_record)
    out_name = in_record.in_name * " earthling"
    return ExampleOutSchemaV1(; out_name)
end
ts = TransformSpecification(ExampleInSchemaV1, ExampleOutSchemaV1, apply_example)

# output
TransformSpecification{ExampleInSchemaV1,ExampleOutSchemaV1}: `apply_example`

transform!(ts, ExampleInSchemaV1(; in_name="greetings"))

# output
ExampleOutSchemaV1: (out_name = "greetings earthling",)

TransformSpecifications.get_step — Method

get_step(dag::NoThrowDAG, name::String) -> DAGStep
get_step(dag::NoThrowDAG, step_index::Int) -> DAGStep

Return DAGStep with name or step_index.

TransformSpecifications.input_specification — Function

input_specification(ts::AbstractTransformSpecification)

Return specification accepted as input to ts.

TransformSpecifications.input_specification — Method

input_specification(dag::NoThrowDAG)

Return input_specification of first step in dag, which is the input specification of the entire DAG.

See also: output_specification, NoThrowDAG

TransformSpecifications.is_identity_no_throw_transform — Method

is_identity_no_throw_transform(ntt::NoThrowTransform) -> Bool

Check if ntt meets the definition of an identity NoThrowTransform, namely, output_specification(ntt) == NoThrowTransform{input_specification(ntt)} and transform function is identity_no_throw_result.

TransformSpecifications.nothrow_succeeded — Method

nothrow_succeeded(result::NoThrowResult) -> Bool

Return true if result indicates successful completion, i.e. if result.violations is empty.

See also: NoThrowResult

TransformSpecifications.output_specification — Function

output_specification(ts::AbstractTransformSpecification)

Return specification of return value of ts.

TransformSpecifications.output_specification — Method

output_specification(dag::NoThrowDAG) -> Type{<:Legolas.AbstractRecord}

Return output_specification of last step in dag, which is the output specification of the entire DAG.

See also: input_specification, NoThrowDAG

TransformSpecifications.transform! — Function

transform!(ts::AbstractTransformSpecification, input)

Return result of applying ts to an input of type input_specification(ts), where result is an output_specification(ts). May mutate input.

See also: transform

TransformSpecifications.transform! — Method

transform!(dag::NoThrowDAG, input)

Return NoThrowResult of sequentially transform!ing all dag.step_transforms, after passing input to the first step.

Before each step, that step's input_assembler is called on the results of all previous processing steps; this constructor generates input that conforms to the step's input_specification.

TransformSpecifications.transform! — Method

transform!(ntt::NoThrowTransform, input)

Return NoThrowResult of applying ntt.transform_spec.transform_fn to input. Transform will fail (i.e., return a NoThrowResult{Missing} if:

input does not conform to input_specification(ntt), i.e., convert_spec(input_specification(ntt), input) throws an error

ntt.transform_spec.transform_fn returns a NoThrowResult{Missing} when applied to the interpreted input,

ntt.transform_spec.transform_fn errors when applied to the interpreted input, or

the output generated by ntt.transform_spec.transform_fn is not a Union{NoThrowResult{Missing},output_specification(ntt)}

In any of these failure cases, this function will not throw, but instead will return the cause of failure in the output violations field.

Note

For debugging purposes, it may be helpful to bypass the "no-throw" feature and so as to have access to a callstack. To do this, use transform_unwrapped! in place of transform!.

See also: convert_spec

TransformSpecifications.transform! — Method

transform!(ts::TransformSpecification, input)

Return output_specification(ts) by applying ts.transform_fn to input. May error if:

input does not conform to input_specification(ts), i.e., convert_spec(input_specification(ts), input) errors

ts.transform_fn errors when applied to the interpreted input, or

the output generated by ts.transform_fn is not a output_specification(ts)

For a non-erroring alternative, see NoThrowTransform.

See also: convert_spec

TransformSpecifications.transform — Method

transform(ts::AbstractTransformSpecification, input)

Return result of applying ts to an input of type input_specification(ts), where result is an output_specification(ts). May not mutate input.

See also: transform!

TransformSpecifications.transform_unwrapped! — Method

transform_unwrapped!(ntt::NoThrowTransform, input)

Apply transform! on inner ntt.transform_spec, such that the resultant output will be of type output_specification(ntt.transform_spec) rather than a NoThrowResult, any failure will result in throwing an error. Utility for debugging NoThrowTransforms.

See also: transform_unwrapped

TransformSpecifications.transform_unwrapped — Method

transform_unwrapped(ntt::NoThrowTransform, input)

Non-mutating implmementation of transform_unwrapped!; applies transform(ntt.transform_spec, input).

Non-exported functions and types

TransformSpecifications._validate_input_assembler — Method

_validate_input_assembler(dag::NoThrowDAG, input_assembler::TransformSpecification)
_validate_input_assembler(dag::NoThrowDAG, ::Nothing)

Confirm that an input_assembler, when called on all upstream outputs generated by dag, has access to all fields it needs to construct its input.

This validation assumes that the dag being validated against contains only outputs from upstream steps, an assumption that will be true on DAG construction (where this validation is called from). If called at other times, validation may succeed even though the input_assembler will fail when called in situ.

TransformSpecifications.convert_spec — Method

convert_spec(::Type{T}, input::T) where {T}
convert_spec(::Type{T}, input::T) where {T<:Legolas.AbstractRecord}
convert_spec(spec::Type{<:Legolas.AbstractRecord}, input)
convert_spec(spec, input)

Return input interpreted as type T: is same as identity function if input is already of type T; otherwise, attempts to construct or Base.converts the the output type from the input. Will throw if conversion fails or is otherwise undefined.

See also: transform!

TransformSpecifications.field_dict — Method

field_dict(type::Type{<:NoThrowResult})
field_dict(type)

Return a Dict where keys are fieldnames(type) and the value of each key is that field's own type. Constructed by calling field_dict_value on each input field's type.

When type is a NoThrowResult{T}, generate mapping based on unwrapped type T.

To recurse into a specific type MyType, implement

TransformSpecification.field_dict_value(t::Type{MyType}) = field_dict(t)

Warning

Use caution when implementing a field_dict_value for any type that isn't explicitly impossible to lead to recursion, as otherwise a stack overflow may occur.

TransformSpecifications.identity_no_throw_result — Method

identity_no_throw_result(result) -> NoThrowResult

Return NoThrowResult{T} where T=typeof(result)

TransformSpecifications.input_assembler — Method

input_assembler(conversion_fn) -> TransformSpecification{Dict{String,Any}, NamedTuple}

Special transform used to convert the outputs of upstream steps in a NoThrowDAG into a NamedTuple that can be converted into that type's input specification.

conversion_fn must be a function that

takes as input a Dictionary with keys that are the names of upstream steps, where the value of each of these keys is the output of that upstreamstep, as specified by `outputspecification(upstream_step)`.

returns a NamedTuple that can be converted, via convert_spec, to the specification of an AbstractTransformSpecification that it is paired with in a DAGStep.

Note that the current implementation is a stopgap for a better-defined implementation defined in https://github.com/beacon-biosignals/TransformSpecifications.jl/issues/8

TransformSpecifications.is_input_assembler — Method

is_input_assembler(ts::AbstractTransformSpecification) -> Bool

Confirm that ts is an input_assembler.